requirements.txt
pandas
shillelagh[all]
pypng
pyqrcode
twilio